+++ internal/domain/event/service.go
@@ -2,9 +2,9 @@ package event
 
 type Service interface {
 	FindAll() (*[]Event, error)
-	FindOne(id int) (*Event, error)
+	FindOne(id uint64) (*Event, error)
 	FindByCoords(float64, float64, float64) (*[]Event, error)
 	PostOne([]byte) error
-	UpdateOne(int, []byte) error
-	DeleteOne(int) error
+	UpdateOne(uint64, []byte) error
+	DeleteOne(uint64) error
 }
